App/requirements.txt
flask
torch
torchvision
numpy
nibabel
pillow
